pyproject.toml
[project]
name = "minecraft-dashboard"
version = "0.1.0"
description = "Near real-time dashboard for vanilla Minecraft server competitions"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery>=3.25",
    "streamlit>=1.38",
    "plotly>=5.24",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "streamlit-autorefresh>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
mc-collect = "collector.main:main"
mc-dashboard = "dashboard.app:main"

[tool.setuptools.packages.find]
include = ["collector*", "dashboard*", "schemas*", "config*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
